Split ComponentUID into a separate module; rework cuid storage and parser
