+++ pyomo/pysp/scenariotree/__init__.py
@@ -17,7 +17,6 @@ from pyomo.core import (value, minimize, maximize,
                         Var, Expression, Block,
                         Objective, SOSConstraint,
                         ComponentUID)
-from pyomo.core.base.block import generate_cuid_names
 from pyomo.core.base.sos import _SOSConstraintData
 from pyomo.repn import generate_standard_repn
 from pyomo.pysp.phutils import (BasicSymbolMap,
