+++ pyomo/pysp/scenariotree/tree_structure.py
@@ -30,7 +30,6 @@ from pyomo.core import (value, minimize, maximize,
                         Var, Expression, Block,
                         Objective, SOSConstraint,
                         ComponentUID)
-from pyomo.core.base.block import generate_cuid_names
 from pyomo.core.base.sos import _SOSConstraintData
 from pyomo.repn import generate_standard_repn
 from pyomo.pysp.phutils import (BasicSymbolMap,
@@ -50,7 +49,8 @@ class _CUIDLabeler(object):
         self._cuid_map = ComponentMap()
 
     def update_cache(self, block):
-        self._cuid_map.update(generate_cuid_names(block))
+        self._cuid_map.update(
+            ComponentUID.generate_cuid_string_map(block, repr_version=1))
 
     def clear_cache(self):
         self._cuid_map = {}
@@ -59,7 +59,7 @@ class _CUIDLabeler(object):
         if obj in self._cuid_map:
             return self._cuid_map[obj]
         else:
-            cuid = repr(ComponentUID(obj))
+            cuid = ComponentUID(obj).get_repr(version=1)
             self._cuid_map[obj] = cuid
             return cuid
